requirements.txt
amqp==2.3.2
beautifulsoup4==4.6.3
billiard==3.5.0.4
bs4==0.0.1
celery==4.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Classy==0.6.10
Flask-SQLAlchemy==2.3.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.1.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycparser==2.19
pytz==2018.7
PyYAML==3.13
rabbitmq==0.2.0
redis==2.10.6
requests==2.20.1
SQLAlchemy==1.2.14
tqdm==4.28.1
urllib3==1.24.1
vine==1.1.4
Werkzeug==0.14.1
